backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pyjwt==2.8.0

# Configuration & Environment
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic[email]==2.5.3

# Caching
redis==5.0.1
aioredis==2.0.1

# File Handling
python-magic==0.4.27
pillow==10.2.0

# Utilities
python-slugify==8.0.1
email-validator==2.1.0
aiofiles==23.2.1
httpx==0.26.0

# Additional Auth/Security
cryptography==42.0.0
pyotp==2.9.0
qrcode[pil]==7.4.2
authlib==1.3.0
itsdangerous==2.1.2

# Payment Gateways
stripe==7.10.0